test-requirements.txt
mock==0.8.0
nose==1.1.2
nosexcover==1.0.7
pylint==0.25.1
lxml==2.3.3
blinker==1.2
coverage==3.5.1
flask-testing
nose-exclude==0.1.8
